Updates cli flags, adds to readme
