+++ Makefile
@@ -25,5 +25,8 @@ license:
 	./scripts/add_license.sh
 
 build:
-	cd cmd/ && go build -o ../build/retrieval-client ./...
+	cd cmd/retrieval-client && go build -o ../build/retrieval-client .
+
+install:
+	cd cmd/retrieval-client && go install
